requirements.txt
annotated-types==0.7.0
anyio==4.4.0
baidu-aip==4.16.13
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.30.3
opencv-python-headless==4.9.0.80
pillow==10.3.0
PyAudio==0.2.14
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
requests==2.32.2
sniffio==1.3.1
SpeechRecognition==3.10.4
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
